cli split: Make `jj split --parallel` set @ to the first revision
`jj split --parallel` now sets the working copy revision (`@`) to the first
revision created by the split instead of the second. As a result, `@` remains
on the same change id before and after the split.

There is no good reason to choose the second commit as the working copy commit
after a --parallel split. The only reason it works that way today is because
doing that didn't require any code changes compared to the non parallel split.

Martin proposed this change in the review for https://github.com/jj-vcs/jj/pull/5618.