docs: add explanation of how to check MSRV locally
There have been a few attempts to add a `rust-toolchain.toml` file to
pin developers' rust version. The most recent attempt was [reverted].

The current opinion ([comment]) is that using a newer compiler
version is desired, and MSRV violations should be caught by
`clippy::incompatible_msrv`. This commit writes that opinion in the
`docs/contributing.md`.

[reverted]: https://github.com/jj-vcs/jj/pull/5090

[comment]: https://github.com/jj-vcs/jj/pull/1913#issuecomment-2037636332