cli split: Implement --parallel by rebasing the second commit
This implements --parallel by creating the second commit as usual, and then
rebasing it onto the parents of the target commit.

I'm not sure if this is more or less efficient. I suppose it might be less
efficient since it writes the second commit twice instead of once, but I'm not
sure if that matters in practice.

This approach should be simpler if we end up with arbitrary destinations for
the split commit via the proposed -A, -B, -d flags since rebase will compute
the new tree for us.