cli: rename `--restore-descendants` to `--preserve-descendant-content` AKA `--pdc`
Fixes https://github.com/jj-vcs/jj/discussions/5951

There have also been discussions on Discord on and off; people seem to
agree that `--preserve-descendant-content` is clear if verbose. I hope
that a short alias might address that (though I don't intent to make a
habit of having many cryptic short aliases), though we could also decide
to skip it for now and see whether it's really needed.

Some other options considered:

- `keep-descendant-content`. In theory, I like it better, but it'd be
  confusingly similar to `jj squash --keep-emptied` once `jj squash
  --keep-descendant-content` was implemented, and would also conflict
  with its tab completion.
- `--preserve-descendant-contents` with `s` at the end.
- The short version could be `-P`, but that would be confusable with
  `--preserve-content` for `jj rebase`, discussed below. Also, currently
  single-capital-letter flags are usually about lifting some
  restriction, e.g. `-B` for `--allow-backwards`.
- `verbatim-descendants`
- `reparent-descendants`

The last two look a bit cryptic, depending on the person.

Note that `jj rebase` will also need a `--preserve-content` flag for
preserving the content of the commit actually being rebased. I'm not
sure whether that should also be `--pc` or not.