config-schema: fix schema default for `revset-aliases."immutable_heads()"`
The default value became definitely outdated when "branches" were renamed
to "bookmarks" in 0.20 (and broke in 0.28).

The actual default of `revset-aliases."immutable_heads()"` is in fact
`builtin_immutable_heads()`, now. However, if one were to want override
to default `immutable_heads()`, `builtin_immutable_heads()` would likely
not be a helpful starting point. Therefore, it may make sense to keep
this "resolved" default value in the schema.