config-schema: validate schema defaults according to schema
While the existing test checks that the schema defaults are consistent
with the output of `jj config get` (with default config), it would not
find type errors like the ones fixed in e9da94e.

This add another test case which validates the synthetic TOML containing
the default values according to the schema against the schema itself.