merge-tools builtin: add regression test for directory replaced by file
This is a manual regression test for issue #5189. When this change is
replaced onto `v0.28.2`, the test produces a panic, as described in the
issue.

As of #6411, the panic no longer happens. However, the test still fails
and rightfully so: When following the reproduction and not selecting any
change in `jj split`, the split-off (first) commit will still record the
deletion of `folder/.keep` but not the creation of the file `folder`.